prototype sharding in the Array class
